removed header styling
