+++ src/components/Modals/PolicyModal/AttendancePolicyModal.test.js
@@ -1,7 +1,8 @@
-import React from "react"
-import { AttendancePolicyModal } from "./AttendancePolicyModal"
-import { render, screen } from "@testing-library/react"
+import { AttendancePolicyModal } from "../../Modals/PolicyModal/AttendancePolicyModal"
 import { BrowserRouter } from "react-router-dom"
+import React from "react"
+import { render, screen } from "@testing-library/react"
+
 
 describe("<PolicyModal />", () => {
   it("displays PolicyModal", () => {
