+++ src/pages/CodeOfConduct/CodeOfConduct.js
@@ -13,7 +13,7 @@ const CodeOfConduct = () => {
         <p className="my-3 max-w-readable">
                 The administration at LEARN academy, endeavors to build a brave, inclusive space—online, in the classroom, and at our events. A primary assumption of a brave space is that everyone speaks with the positive intent of seeking greater knowledge and understanding of differing perspectives and experiences. To promote a spirit of inclusivity, we ask our instructors, students, mentors, and guests to follow a few foundational principles while sharing our space. These principles are embodied by, but not limited to, the following statements:
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 We are Welcoming, Friendly, and Patient
         </h2>
         <p className="my-3 max-w-readable">
@@ -22,37 +22,37 @@ const CodeOfConduct = () => {
         <p className="my-3 max-w-readable">
                 It is crucial that every individual knows that LEARN academy is an inclusive, respectful, community-oriented learning environment. Each individual in the LEARN academy community has a part to play in building and maintaining these principles.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 We are Considerate
         </h2>
         <p className="my-3 max-w-readable">
                 Individuals work and decisions affect others. Whether writing code, giving a presentation or participating in a discussion, individuals’ actions have consequences. Individuals should take those consequences into account when making decisions.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 We are Respectful
         </h2>
         <p className="my-3 max-w-readable">
                 We can all experience frustration now and then; coding is hard work and can be exasperating. How can we ensure that frustration does not get in the way of positive, productive relationships with others? Give yourself grace if you feel frustrated with coding, another member of our community, or with yourself. LEARN staff is available at any time for individuals who need to talk through frustration or need help facilitating communication with another member of the community.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 We are Mindful of Vocabulary
         </h2>
         <p className="my-3 max-w-readable">
                 Remember that sexist, racist, and other exclusionary jokes or media content can be offensive to others. Consider the content of the music you play, the videos you stream, the images you share, and the language you use in our learning space. Do they contain words or images that some may find offensive or denigrating? Be kind to others, and do not insult or put down another person. Mindfulness often means reevaluating what we consider permissible.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 We use our Skills for Good
         </h2>
         <p className="my-3 max-w-readable">
                 Computers are a revolutionary communication tool, which connects us like never before. With this connectivity comes a responsibility to respect others’ property and rights on the internet. Our duty goes beyond following the law, which is a must. At LEARN academy, we use our facility, resources, and skills to do “good” in the community. We help others make the internet safe as well, by following best practices of netizenship, such as reporting issues found on the internet or within our community in a manner consistent with LEARN academy’s values.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 Reporting Inappropriate Behavior
         </h2>
         <p className="my-3 max-w-readable">
                 We care about individuals’ experiences while they are with us at LEARN academy. Chelsea and Sarah are trained in conflict resolution and diversity and inclusion issues, and they are available to talk with individuals about any topic. If an individual’s experience as a member of the LEARN academy community has made them uncomfortable, feel unwelcome, or has in any other way not met their expectations of a professional learning environment, we are here to listen and to protect the brave space.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
                 At LEARN academy, we promise:
         </h2>
         <p className="my-3 max-w-readable">
@@ -74,7 +74,7 @@ const CodeOfConduct = () => {
                 Sarah Proctor: <a className="text-learnpurple hover:text-learngreen active:text-learngreen" href="mailto:Sarah@learnacademy.org">Sarah@learnacademy.org</a>
         </p>
         <br></br>
-        <h3 className="max-w-readable text-lg text-learnblack">
+        <h3 className="text-lg text-learnblack">
                 Additional Resources:
         </h3>
         <ol className="ml-8 list-disc">
