+++ src/pages/PrivacyPolicy/PrivacyPolicy.js
@@ -7,7 +7,7 @@ const PrivacyPolicy = () => {
         <h1 className="mt-4 mb-12 text-4xl font-semibold text-learnblack">
     Privacy Policy
         </h1>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Your privacy is critically important to us.
         </h2>
         <p className="my-3 max-w-readable">
@@ -23,7 +23,7 @@ const PrivacyPolicy = () => {
         <p className="my-3 max-w-readable">
     This Privacy Policy, together with the Terms and conditions posted on our Website, set forth the general rules and policies governing your use of our Website. Depending on your activities when visiting our Website, you may be required to agree to additional terms and conditions.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Website Visitors
         </h2>
         <p className="my-3 max-w-readable">
@@ -32,25 +32,25 @@ const PrivacyPolicy = () => {
         <p className="my-3 max-w-readable">
     LEARN academy also collects potentially personally-identifying information like Internet Protocol (IP) addresses for logged in users and for users leaving comments on https://www.learnacademy.org blog posts. LEARN academy only discloses logged in user and commenter IP addresses under the same circumstances that it uses and discloses personally-identifying information as described below.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Gathering of Personally-Identifying Information
         </h2>
         <p className="my-3 max-w-readable">
     Certain visitors to LEARN academy&apos;s websites choose to interact with LEARN academy in ways that require LEARN academy to gather personally-identifying information. The amount and type of information that LEARN academy gathers depends on the nature of the interaction. For example, we ask visitors who sign up for a blog at https://www.learnacademy.org to provide a username and email address.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Security
         </h2>
         <p className="my-3 max-w-readable">
     The security of your Personal Information is important to us, but remember that no method of transmission over the Internet, or method of electronic storage is 100% secure. While we strive to use commercially acceptable means to protect your Personal Information, we cannot guarantee its absolute security. 
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Advertisements
         </h2>
         <p className="my-3 max-w-readable">
     Ads appearing on our website may be delivered to users by advertising partners, who may set cookies. These cookies allow the ad server to recognize your computer each time they send you an online advertisement to compile information about you or others who use your computer. This information allows ad networks to, among other things, deliver targeted advertisements that they believe will be of most interest to you. This Privacy Policy covers the use of cookies by LEARN academy and does not cover the use of cookies by any advertisers.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Links to External Sites
         </h2>
         <p className="my-3  max-w-readable">
@@ -59,13 +59,13 @@ const PrivacyPolicy = () => {
         <p className="my-3 max-w-readable">
     We have no control over, and assume no responsibility for the content, privacy policies or practices of any third party sites, products or services.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Aggregated Statistics
         </h2>
         <p className="my-3 max-w-readable">
     LEARN academy may collect statistics about the behavior of visitors to its website. LEARN academy may display this information publicly or provide it to others. However, LEARN academy does not disclose your personally-identifying information.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Cookies
         </h2>
         <p className="my-3 max-w-readable">
@@ -75,13 +75,13 @@ const PrivacyPolicy = () => {
     
     By continuing to navigate our website without changing your cookie settings, you hereby acknowledge and agree to LEARN academy&apos;s use of cookies.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Privacy Policy Changes
         </h2>
         <p className="my-3 max-w-readable">
         Although most changes are likely to be minor, LEARN academy may change its Privacy Policy from time to time, and in LEARN academy&apos;s sole discretion. LEARN academy encourages visitors to frequently check this page for any changes to its Privacy Policy. Your continued use of this site after any change in this Privacy Policy will constitute your acceptance of such change.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
     Credit & Contact Information
         </h2>
         <p className="my-3 max-w-readable">
