+++ src/pages/TermsOfUse/TermsOfUse.js
@@ -7,7 +7,7 @@ const TermsOfUse = () => {
         <h1 className="mt-4 mb-12 text-4xl font-semibold text-learnblack">
           Terms & Conditions
         </h1>
-        <h2 className="max-w-readable text-2xl text-learnblack">Welcome to LEARN academy</h2>
+        <h2 className="text-2xl text-learnblack">Welcome to LEARN academy</h2>
         <p className="my-3 max-w-readable">
           These terms and conditions outline the rules and regulations for the
           use of LEARN academy&apos;s Website.
@@ -48,7 +48,7 @@ const TermsOfUse = () => {
           plural, capitalization and/or he/she or they, are taken as
           interchangeable and therefore as referring to same.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">Cookies</h2>
+        <h2 className="text-2xl text-learnblack">Cookies</h2>
         <p className="my-3 max-w-readable">
           We employ the use of cookies. By using LEARN academy&apos;s website
           you consent to the use of cookies in accordance with LEARN
@@ -61,7 +61,7 @@ const TermsOfUse = () => {
           use for those people visiting. Some of our affiliate/advertising
           partners may also use cookies.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">License</h2>
+        <h2 className="text-2xl text-learnblack">License</h2>
         <p className="my-3 max-w-readable">
           Unless otherwise stated, LEARN academy and/or it&apos;s licensors own
           the intellectual property rights for all material on LEARN academy.
@@ -95,7 +95,7 @@ const TermsOfUse = () => {
             specifically made for redistribution)
           </li>
         </ol>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
           Hyperlinking to our Content
         </h2>
         <p className="my-3 max-w-readable">
@@ -205,13 +205,13 @@ const TermsOfUse = () => {
           No use of LEARN academy&apos;s logo or other artwork will be allowed
           for linking absent a trademark license agreement.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">Iframes</h2>
+        <h2 className="text-2xl text-learnblack">Iframes</h2>
         <p className="my-3 max-w-readable">
           Without prior approval and express written permission, you may not
           create frames around our Web pages or use other techniques that alter
           in any way the visual presentation or appearance of our Web site.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">Reservation of Rights</h2>
+        <h2 className="text-2xl text-learnblack">Reservation of Rights</h2>
         <p className="my-3 max-w-readable">
           We reserve the right at any time and in its sole discretion to request
           that you remove all links or any particular link to our Web site. You
@@ -221,7 +221,7 @@ const TermsOfUse = () => {
           site, you agree to be bound to and abide by these linking terms and
           conditions.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">
+        <h2 className="text-2xl text-learnblack">
           Removal of links from our website
         </h2>
         <p className="my-3 max-w-readable">
@@ -236,7 +236,7 @@ const TermsOfUse = () => {
           commit to ensuring that the website remains available or that the
           material on the website is kept up to date.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">Content Liability</h2>
+        <h2 className="text-2xl text-learnblack">Content Liability</h2>
         <p className="my-3 max-w-readable">
           We shall have no responsibility or liability for any content appearing
           on your Web site. You agree to indemnify and defend us against all
@@ -246,7 +246,7 @@ const TermsOfUse = () => {
           criminal, or which infringes, otherwise violates, or advocates the
           infringement or other violation of, any third party rights.
         </p>
-        <h2 className="max-w-readable text-2xl text-learnblack">Disclaimer</h2>
+        <h2 className="text-2xl text-learnblack">Disclaimer</h2>
         <p className="my-3 max-w-readable">
           To the maximum extent permitted by applicable law, we exclude all
           representations, warranties, and conditions relating to our website
